If we are using --packages and/or --packages-newer and we have to build anyway we must run the 'make config' code in addition to the build dep check since a) the user might actually want to change something, and b) if using -H and there are OPTIONS the build will hang.
To facilitate this turn the actual code to do 'make config' into a function

Testing if an existing package is actually readable is slightly more
useful than just testing if it exists.

If the package does exist, and we're not using --always-fetch then
return 0, both to skip the rest of the function and to make good use
of the return value in the 'second chance fetch' code.