In the test for -PP being alone on the command line I forgot the -dash at the beginning of the second test so it was falsely picking up ports with names like p5-Devel-PPPort.
If you are building more than one port on the command line, and you
use --packages-build and/or --delete-build-only, and one of the ports
on the command line would otherwise qualfiy as a build-only dep, it was
falsely being treated as such. Solve this problem by adding all ports
on the command line to the run_dl_g list in multiport().