Bug fix: ======== PM_BUILD_ONLY_LIST should only be set to pm_bol in the command line parser if we are in the parent.
Feature expansion:
==================
Do a more thorough job of tracking things that should fall into the
--packages-build/--delete-build-only category. I.e., if something is
a run dependency of a port or ports that themselves are only build
dependencies, add it to the build_only_dl_g anyway.

New internal:
=============
Add a uniquify_list function so that internal lists of variables
can be made unique. Start using it for some of the new stuff, and
in clean_build_only_list().

Cleanup:
========
Factor out one line of common code for the rundep list in dep_check()