1. Add code so that if a user has an +IGNOREME file for a port that is a dependency and chooses not to update it we will not falsely claim that it will be udpated in post_config().
This requires moving the writing of num_of_deps and build_l to IPC_SAVE
into safe_exit() so that the modified values will be read by the parent.

2. If a user chose -G don't print the (${dep_of_deps}/${num_of_deps})
in the xterm title since the latter will not be incremented. This has
the pleasant side effect of allowing better indenting of the code.

3. Remove debug code.