Next steps in --index-only support: 1. In usage() show that --index and --index-only are mutually exclusive 2. Conditionalize cd'ing to the portdir in a lot more places 3. Add gen_dep_list() to generate a list of dependencies either by using the old method or from parsing the INDEX file and use it when needed. 4. Add use of the INDEX to figure out the "new" version of the port 5. When using --index-only we can find the expected value of PORTSDIR and LOCALBASE from the INDEX file. For the former this requires moving the parsing of the INDEX up above setting PORTSDIR.
Other changes:
pd and pdb are used too many places to change them all, so introduce tests
for PORTSDIR and PKG_DBDIR being set in the environment so that the user
can set familiar variables on the rc file.

New Feature:
Add a --delete-packages option to delete the package after we're done
installing it.

Bug Fixes:
1. Use pm_cd_pd() more consistently
2. The fetch and pkg_add commands both require $PM_SU_CMD