Call init_packages() in pm_pkg_create() so that if a parallel portmaster cleaned up empty package directories we can still create packages. Also, call init_packages() unconditionally in -s mode. The function has been optimized sufficiently so that concerns about doing duplicate work are no longer valid, and it's safer at this point to just do it.
Add a pm_sv() for printing verbose messages when using sudo, and convert
existing messages to use it. This cleans up the code quite a bit, and will
also help in debugging efforts.

Use pm_make_s (sudo version) for the 'clean' target. In the normal case
this is not necessary since the port is built as the unprivileged user,
however some ports (e.g., thunderbird-3) do 'stuff' in the 'install'
target (which is priviliged) that the unprivileged user can't undo.