Expand on an idea provided by mandree@ to use 'pkg_version -I' to quickly check what ports are out of date, and incorporate that check into the --index option. Feedback from users seems to indicate that this is what they expected the option did anyway, so the minor change to the feature does not seem to be a POLA violation, especially since INDEX support is still relatively new. Add the --index-first option to do what --index used to do in case anyone cares.
Add a bunch more error checking for incompatible command line options.
Add a new cross_idx() to handle checking the 3 --index* options and
consolidate some of the old tests that were all relative to -G.

We want to test the .bz2 version of INDEX to see if it's newer, and
unzip it if it is.