Remove no longer necessary local variable in version()
Add a --no-index-fetch option for people who want to use the
--index* options but do not want portmaster to do the fetching.

Rework how some initial variables are set. Deriving the values via
make works, but is very slow. So try to be more intelligent about
recognizing the standard values for the ones we always need to have.
  * For PORTSDIR and PKG_DBDIR if those variables are not in the
    environment and the standard directories exist, use them.
  * Move the setting of pd (PORTSDIR) up to before the INDEX-related
    items so that we can use ${INDEXDIR:-$pd}
  * The derivation of the INDEX-related stuff wasn't working (since
    we didn't know $pd) and no one has complained, so just use the
    default values from bsd.port.mk unless the user specifies others.
  * If /var/db/ports exists, use it for port_dbdir

In the test to see if we downloaded a new INDEX.bz2 file supply a
default value for index_time so that if the file did not exist
previously the test will not fail.

Indicate that the -[lL] options are not compatible with -FRaefnors,
updates, or installs