1. Use a better, more generic method of signaling an abnormal exit 2. Call safe_exit with a status of 1 from more places.
The specific problem the latter solves is when using -r/f but not -R, and
PM_UPGRADE_DONE_FLAG files exist from a previous run. If an abnormal exit
is not signaled when necessary in this situation the flags will get deleted
even though we want to keep them.

This is likely the cause of some of the vague problem reports I have
received in the past that include complaints of, "I used -r/f and -R
but portmaster still rebuilt certain ports a whole bunch of times."