Improve the regexp for parse_index() by specifying the PORTSDIR path, and not using globals. In the initialization code when using --index* set the path for the ports tree in the INDEX file with a special version of $pd, PM_IPD. We only need to set this to something other than /usr/ports if $pd is not already set, and/or it's set to something other than /usr/ports. In addition to improving the efficiency of parse_index() this has the pleasant side effect of (slightly) improving support for non-standard PORTSDIR.
For -F, at the end of the run break out of the loop if the number of
extant fetches is 0. This doesn't happen often, but there is a race
that can cause it. Printing the message that we're waiting for zero
fetches is silly; but the real problem is sleep'ing for no reason.