When calling trap_exit() in a child process we need to signal the parent(s) that we trap'ed so that the various post-run cleanups and updates can still happen properly.
In a related change, in safe_exit() remove a shortcut from r214541 that
was safe at the time, but isn't any longer because we now have more
stuff happening after a run so we still need to update the parent even
if the child fail'ed or trap'ed.

When we're run by script(1) the parent process is going to be [sh], not
$0, so go back to the less efficient but safer method of safeguarding
the parent in kill_bad_children().

For -a remove a -v line that is redundant now with the new in-line
term_printf() code.

In the new in-line printing code for term_printf() use a simpler
pattern to match $1 that also catches the following ...

In multiport(), update the total number of ports after the first pass
to account for ports given on the command line that turn out to be
dependencies of other ports given on the command line, and print a
summary with the number of parent ports and dependencies prior to the
confirmation prompt.