Revert r237003 since it is needed to restore the terminal title during a long delay between -a ports to update. Add a comment so I don't forget.
Also restore part of r236996, match >> for the in-line term_printf()
code, and add that pattern in post_first_pass() where it's needed to match.
Adjust the summary message in multiport() to suit. Arguably this is a more
elegant way to present the information anyway.

Don't bother trying to delete empty directories in WRKDIRPREFIX if we
are exiting abnormally, chances are there will be a work directory
from the failed port.