Sync orphan flag handling with pkgng logic
Prior to this change, portmaster used to set as non-automatic every
port specified on the command line. Now pkgng inherits orphan flag
even when `pkg install`ing a port explicitly (if it was already
installed, of course), so let's do the same in portmaster: always
inherit the flag when a port is already installed. Only special case
is with -o, where the port is set as non-automatic if one of the two
ports specified on the command line is already like that.