Add messages to make the 'Terminated' messages seem less unexpected
I've verified that the parent process is properly and safely ignored.
This is just cleaning up any background fetch/distinfo background
processes. The only other solution is to redirect stderr to /dev/null
in those jobs, which would hide other real errors.

Hopefully by wrapping the kill/exit in some messages it becomes more
clear that these are expected and not a bug.