Fix previous commit
The patch file contained 2 typoes ("{$" instead of "${").

Add clean phase for dependencies that are not installed but required when
the dependent port is built (e.g. because of references to source files of
that port, or because of static linking against libraries in the stage area
of such a dependency).

Port revision:	3.19_2

Merged from port commit ba1f4c6e0bcf57df430becefc0581498c815103d