Make --clean-packages consider flavors of packages
This commit fixes GutHub issue #55 for flavored ports. Some ports have
package names that are prefixed by e.g. an interpreter version, but do
not support flavors. Such cases are not dealt with by this patch,
since there is no generic way to do so.

The fix does not apply to --clean-packages used with --index or
--index-only, since the INDEX file does not record port flavors.