Fix issue when all updates use existing packages
When using the -P or --package option to portmaster and all updates
can be performed without building at least one port, the $DI_FILES
variable may be unset.

In this situation portmaster stops after the installation of each
port. Entering a CR makes it continue.

This was reported by Vlad Markov with a suggested fix that has been
applied (with a white-space change).